+++ requirements.txt
@@ -1,2 +1,6 @@
+colorama==0.4.4
+commonmark==0.9.1
 greenlet==1.1.1
+Pygments==2.10.0
+rich==10.9.0
 SQLAlchemy==1.4.23
